fix survey calculation on surveyfragment bugs
